+++ js/views/result-attempts-left-view.js
@@ -1,14 +1,14 @@
-import AbstractView from '../abstract-view';
+import AbstractView from './abstract-view';
 
-export default class ResultTimeLeftView extends AbstractView {
+export default class ResultAttemptsLeftView extends AbstractView {
 
   get template() {
     return `
     <section class="main main--result">
       <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
 
-      <h2 class="title">Увы и ах!</h2>
-      <div class="main-stat">Время вышло!<br>Вы не успели отгадать все мелодии</div>
+      <h2 class="title">Какая жалость!</h2>
+      <div class="main-stat">У вас закончились все попытки.<br>Ничего, повезёт в следующий раз!</div>
       <span role="button" tabindex="0" class="main-replay">Попробовать ещё раз</span>
     </section>`;
   }
