+++ js/views/welcome-view.js
@@ -1,4 +1,4 @@
-import AbstractView from '../abstract-view';
+import AbstractView from './abstract-view';
 
 export default class WelcomeView extends AbstractView {
   constructor() {
